bnum
Struct BIntD32

pub struct BIntD32<const N: usize> { /* private fields */ }

Big signed integer type, of fixed size which must be known at compile time.

Digits of the underlying BUintD32 are stored in little endian (least significant digit first). This integer type aims to exactly replicate the behaviours of Rust’s built-in signed integer types: i8, i16, i32, i64, i128 and isize. The const generic parameter N is the number of digits that are stored in the underlying BUintD32.

BIntD32 implements all the arithmetic traits from the core::ops module. The behaviour of the implementation of these traits is the same as for Rust’s primitive integers - i.e. in debug mode it panics on overflow, and in release mode it performs two’s complement wrapping (see https://doc.rust-lang.org/book/ch03-02-data-types.html#integer-overflow). However, an attempt to divide by zero or calculate a remainder with a divisor of zero will always panic, unless the checked_ methods are used, which never panic.

Implementations

impl<const N: usize> BIntD32<N>

Checked arithmetic methods which act on self: self.checked_.... Each method cannot panic and returns an Option<Self>. None is returned when overflow would have occurred or there was an attempt to divide by zero or calculate a remainder with a divisor of zero.

pub const fn checked_add(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_add.

pub const fn checked_add_unsigned(self, rhs: BUintD32<N>) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_add_unsigned.

pub const fn checked_sub(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_sub.

pub const fn checked_sub_unsigned(self, rhs: BUintD32<N>) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_sub_unsigned.

pub fn checked_mul(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_mul.

pub fn checked_div(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_div.

pub fn checked_div_euclid(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_div_euclid.

pub fn checked_rem(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_rem.

pub fn checked_rem_euclid(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_rem_euclid.

pub fn checked_neg(self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_neg.

pub fn checked_shl(self, rhs: u32) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_shl.

pub fn checked_shr(self, rhs: u32) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_shr.

pub fn checked_abs(self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_abs.

pub fn checked_pow(self, pow: u32) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_pow.

pub fn checked_next_multiple_of(self, rhs: Self) -> Option<Self>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_next_multiple_of.

pub fn checked_ilog(self, base: Self) -> Option<u32>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_ilog.

pub fn checked_ilog2(self) -> Option<u32>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_ilog2.

pub fn checked_ilog10(self) -> Option<u32>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.checked_ilog10.

impl<const N: usize> BIntD32<N>

Associated constants for this type.

pub const MIN: Self = _

The minimum value that this type can represent.

Examples

use bnum::types::I512;

assert_eq!(!I512::MIN, I512::MAX);

pub const MAX: Self = _

The maximum value that this type can represent.

Examples

use bnum::types::I512;

assert_eq!(I512::MAX.wrapping_add(I512::ONE), I512::MIN);

pub const BITS: u32 = BUintD32<N>::BITS

The total number of bits that this type contains.

Examples

use bnum::types::I512;

assert_eq!(I512::BITS, 512);

pub const BYTES: u32 = BUintD32<N>::BYTES

The total number of bytes that this type contains.

Examples

use bnum::types::I512;

assert_eq!(I512::BYTES, 512 / 8);

pub const ZERO: Self = _

The value of 0 represented by this type.

Examples

use bnum::types::I512;

assert_eq!(I512::ZERO, I512::from(0u8));

pub const ONE: Self = _

The value of 1 represented by this type.

Examples

use bnum::types::I512;

assert_eq!(I512::ONE, I512::from(1u8));

pub const TWO: Self = _

The value of 2 represented by this type.

pub const THREE: Self = _

The value of 3 represented by this type.

pub const FOUR: Self = _

The value of 4 represented by this type.

pub const FIVE: Self = _

The value of 5 represented by this type.

pub const SIX: Self = _

The value of 6 represented by this type.

pub const SEVEN: Self = _

The value of 7 represented by this type.

pub const EIGHT: Self = _

The value of 8 represented by this type.

pub const NINE: Self = _

The value of 9 represented by this type.

pub const TEN: Self = _

The value of 10 represented by this type.

pub const NEG_ONE: Self = _

The value of -1 represented by this type.

pub const NEG_TWO: Self = _

The value of -2 represented by this type.

pub const NEG_THREE: Self = _

The value of -3 represented by this type.

pub const NEG_FOUR: Self = _

The value of -4 represented by this type.

pub const NEG_FIVE: Self = _

The value of -5 represented by this type.

pub const NEG_SIX: Self = _

The value of -6 represented by this type.

pub const NEG_SEVEN: Self = _

The value of -7 represented by this type.

pub const NEG_EIGHT: Self = _

The value of -8 represented by this type.

pub const NEG_NINE: Self = _

The value of -9 represented by this type.

pub const NEG_TEN: Self = _

The value of -10 represented by this type.

impl<const N: usize> BIntD32<N>

Methods which convert a BIntD32 to and from data stored in different endianness.

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.from_be.

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.from_le.

pub const fn to_be(self) -> Self

Converts self from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.to_be.

pub const fn to_le(self) -> Self

Converts self from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.to_le.

pub fn from_be_slice(slice: &[u8]) -> Option<Self>

Create an integer value from a slice of bytes in big endian. The value is wrapped in an Option as the integer represented by the slice of bytes may represent an integer too large to be represented by the type.

If the length of the slice is shorter than Self::BYTES, the slice is padded with zeros or ones at the start so that it’s length equals Self::BYTES. It is padded with ones if the bytes represent a negative integer, otherwise it is padded with zeros.

If the length of the slice is longer than Self::BYTES, None will be returned, unless the bytes represent a non-negative integer and leading zeros from the slice can be removed until the length of the slice equals Self::BYTES, or if the bytes represent a negative integer and leading ones from the slice can be removed until the length of the slice equals Self::BYTES.

For examples, see the from_be_slice method documentation for BUintD32.

pub fn from_le_slice(slice: &[u8]) -> Option<Self>

Creates an integer value from a slice of bytes in little endian. The value is wrapped in an Option as the bytes may represent an integer too large to be represented by the type.

If the length of the slice is shorter than Self::BYTES, the slice is padded with zeros or ones at the end so that it’s length equals Self::BYTES. It is padded with ones if the bytes represent a negative integer, otherwise it is padded with zeros.

If the length of the slice is longer than Self::BYTES, None will be returned, unless the bytes represent a non-negative integer and trailing zeros from the slice can be removed until the length of the slice equals Self::BYTES, or if the bytes represent a negative integer and trailing ones from the slice can be removed until the length of the slice equals Self::BYTES.

For examples, see the from_le_slice method documentation for BUintD32.

impl<const N: usize> BIntD32<N>

Overflowing arithmetic methods which act on self: self.overflowing_.... Each method returns a tuple of type (Self, bool) where the first item of the tuple is the result of the calculation truncated to the number of bits of self, and the second item is a boolean which indicates whether overflow occurred (i.e. if the number of bits of the result of the calculation exceeded the number of bits of self).

pub const fn overflowing_add(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_add.

pub const fn overflowing_add_unsigned(self, rhs: BUintD32<N>) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_add_unsigned.

pub const fn overflowing_sub(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_sub.

pub const fn overflowing_sub_unsigned(self, rhs: BUintD32<N>) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_sub_unsigned.

pub fn overflowing_mul(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_mul.

pub fn overflowing_div(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_div.

pub fn overflowing_div_euclid(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_div_euclid.

pub fn overflowing_rem(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_rem.

pub fn overflowing_rem_euclid(self, rhs: Self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_rem_euclid.

pub fn overflowing_neg(self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_neg.

pub fn overflowing_shl(self, rhs: u32) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_shl.

pub fn overflowing_shr(self, rhs: u32) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_shr.

pub fn overflowing_abs(self) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_abs.

pub fn overflowing_pow(self, pow: u32) -> (Self, bool)

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.overflowing_pow.

impl<const N: usize> BIntD32<N>

Methods which convert a BIntD32 between different types in a given radix (base).

pub fn parse_bytes(buf: &[u8], radix: u32) -> Option<Self>

Converts a byte slice in a given base to an integer. The input slice must contain ascii/utf8 characters in [0-9a-zA-Z].

This function is equivalent to the from_str_radix function for a string slice equivalent to the byte slice and the same radix.

Returns None if the conversion of the byte slice to string slice fails or if a digit is larger than or equal to the given radix, otherwise the integer is wrapped in Some.

pub fn from_radix_be(buf: &[u8], radix: u32) -> Option<Self>

Converts a slice of big-endian digits in the given radix to an integer. The digits are first converted to an unsigned integer, then this is transmuted to a signed integer. Each u8 of the slice is interpreted as one digit of base radix of the number, so this function will return None if any digit is greater than or equal to radix, otherwise the integer is wrapped in Some.

For examples, see the from_radix_be method documentation for BUintD32.

pub fn from_radix_le(buf: &[u8], radix: u32) -> Option<Self>

Converts a slice of big-endian digits in the given radix to an integer. The digits are first converted to an unsigned integer, then this is transmuted to a signed integer. Each u8 of the slice is interpreted as one digit of base radix of the number, so this function will return None if any digit is greater than or equal to radix, otherwise the integer is wrapped in Some.

For examples, see the from_radix_le method documentation for BUintD32.

pub fn from_str_radix(src: &str, radix: u32) -> Result<Self, ParseIntError>

Converts a string slice in a given base to an integer.

The string is expected to be an optional + or - sign followed by digits. Leading and trailing whitespace represent an error. Digits are a subset of these characters, depending on radix:

0-9

a-z

A-Z

Panics

This function panics if radix is not in the range from 2 to 36 inclusive.

For examples, see the from_str_radix method documentation for BUintD32.

pub fn to_str_radix(&self, radix: u32) -> String

Returns the integer as a string in the given radix.

Panics

This function panics if radix is not in the range from 2 to 36 inclusive.

For examples, see the to_str_radix method documentation for BUintD32.

pub fn to_radix_be(&self, radix: u32) -> Vec<u8>

Returns the integer’s underlying representation as an unsigned integer in the given base in big-endian digit order.

Panics

This function panics if radix is not in the range from 2 to 256 inclusive.

For examples, see the to_radix_be method documentation for BUintD32

pub fn to_radix_le(&self, radix: u32) -> Vec<u8>

Returns the integer’s underlying representation as an unsigned integer in the given base in little-endian digit order.

Panics

This function panics if radix is not in the range from 2 to 256 inclusive.

For examples, see the to_radix_le method documentation for BUintD32.

impl<const N: usize> BIntD32<N>

Saturating arithmetic methods which act on self: self.saturating_.... For each method, if overflow or underflow occurs, the largest or smallest value that can be represented by Self is returned instead.

pub const fn saturating_add(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_add.

pub const fn saturating_add_unsigned(self, rhs: BUintD32<N>) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_add_unsigned.

pub const fn saturating_sub(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_sub.

pub const fn saturating_sub_unsigned(self, rhs: BUintD32<N>) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_sub_unsigned.

pub fn saturating_mul(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_mul.

pub fn saturating_div(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_div.

pub fn saturating_neg(self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_neg.

pub fn saturating_abs(self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_abs.

pub fn saturating_pow(self, exp: u32) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.saturating_pow.

impl<const N: usize> BIntD32<N>

Unchecked arithmetic methods which act on self: self.unchecked_.... Each method results in undefined behavior if overflow/underflow occurs, i.e. when the checked equivalent would return None.

pub unsafe fn unchecked_add(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.unchecked_add.

pub unsafe fn unchecked_sub(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.unchecked_sub.

pub unsafe fn unchecked_mul(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.unchecked_mul.

pub unsafe fn unchecked_shl(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.unchecked_shl.

pub unsafe fn unchecked_shr(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.unchecked_shr.

impl<const N: usize> BIntD32<N>

Wrapping arithmetic methods which act on self: self.wrapping_.... Each method returns of the calculation truncated to the number of bits of self, i.e. they each return the first item in the tuple returned by their overflowing equivalent.

pub const fn wrapping_add(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_add.

pub const fn wrapping_add_unsigned(self, rhs: BUintD32<N>) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_add_unsigned.

pub const fn wrapping_sub(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_sub.

pub const fn wrapping_sub_unsigned(self, rhs: BUintD32<N>) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_sub_unsigned.

pub const fn wrapping_mul(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_mul.

pub fn wrapping_div(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_div.

pub fn wrapping_div_euclid(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_div_euclid.

pub fn wrapping_rem(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_rem.

pub fn wrapping_rem_euclid(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_rem_euclid.

pub fn wrapping_neg(self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_neg.

pub fn wrapping_shl(self, rhs: u32) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_shl.

pub fn wrapping_shr(self, rhs: u32) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_shr.

pub fn wrapping_abs(self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_abs.

pub const fn wrapping_pow(self, pow: u32) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.wrapping_pow.

impl<const N: usize> BIntD32<N>

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.count_ones.

Examples

use bnum::types::I256;

let n = I256::from(0b010111101010000u16);
assert_eq!(n.count_ones(), 7);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.count_zeros.

Examples

use bnum::types::I256;

assert_eq!((!I256::ZERO).count_zeros(), 0);

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.leading_zeros.

Examples

use bnum::types::I256;

let n = I256::ZERO;
assert_eq!(n.leading_zeros(), 256);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.trailing_zeros.

Examples

use bnum::types::I256;

let n = (!I256::ZERO) << 6u32;
assert_eq!(n.trailing_zeros(), 6);

pub const fn leading_ones(self) -> u32

Returns the number of leading ones in the binary representation of self.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.leading_ones.

Examples

use bnum::types::I256;

let n = !I256::ZERO;
assert_eq!(n.leading_ones(), 256);

pub const fn trailing_ones(self) -> u32

Returns the number of trailing ones in the binary representation of self.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.trailing_ones.

Examples

use bnum::types::I256;

let n = I256::from(0b1000101011011u16);
assert_eq!(n.trailing_ones(), 2);

pub fn rotate_left(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isn’t the same operation as the << shifting operator!

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.rotate_left.

pub fn rotate_right(self, n: u32) -> Self

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isn’t the same operation as the >> shifting operator!

self.rotate_right(n) is equivalent to self.rotate_left(Self::BITS - n).

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.rotate_right.

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.swap_bytes.

Examples

use bnum::types::I256;

let n = I256::from(0x12345678901234567890123456789012i128);
assert_eq!(n.swap_bytes().swap_bytes(), n);

pub const fn reverse_bits(self) -> Self

Reverses the order of bits in the integer. The least significant bit becomes the most significant bit, second least-significant bit becomes second most-significant bit, etc.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.reverse_bits.

Examples

use bnum::types::I256;

let n = I256::from(0x12345678901234567890123456789012i128);
assert_eq!(n.reverse_bits().reverse_bits(), n);

pub fn unsigned_abs(self) -> BUintD32<N>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.unsigned_abs.

pub fn pow(self, exp: u32) -> Self

Raises self to the power of exp, using exponentiation by squaring.

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.pow.

Examples

use bnum::types::I256;

let n = I256::from(3u8);
assert_eq!(n.pow(5), (243u8).into());

pub fn div_euclid(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.div_euclid.

pub fn rem_euclid(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.rem_euclid.

pub fn abs(self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.abs.

pub const fn signum(self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.signum.

pub const fn is_positive(self) -> bool

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.is_positive.

pub const fn is_negative(self) -> bool

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.is_negative.

pub const fn is_power_of_two(self) -> bool

Returns true if and only if self == 2^k for some integer k.

Examples

use bnum::types::I256;

let n = I256::from(1i16 << 12);
assert!(n.is_power_of_two());
let m = I256::from(90i8);
assert!(!m.is_power_of_two());
assert!(!((-n).is_power_of_two()));

pub fn ilog(self, base: Self) -> u32

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.ilog.

pub fn ilog2(self) -> u32

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.ilog2.

pub fn ilog10(self) -> u32

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.ilog10.

pub fn abs_diff(self, other: Self) -> BUintD32<N>

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.abs_diff.

pub fn next_multiple_of(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.next_multiple_of.

pub fn div_floor(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.div_floor.

pub fn div_ceil(self, rhs: Self) -> Self

See also: https://doc.rust-lang.org/std/primitive.i64.html#method.div_ceil.

impl<const N: usize> BIntD32<N>

pub const fn bits(&self) -> u32

Returns the smallest number of bits necessary to represent self.

This is equal to the size of the type in bits minus the leading zeros of self.

Examples

use bnum::types::I256;

assert_eq!(I256::from(0b1111001010100u16).bits(), 13);
assert_eq!(I256::ZERO.bits(), 0);

pub const fn bit(&self, b: u32) -> bool

Returns a boolean representing the bit in the given position (true if the bit is set). The least significant bit is at index 0, the most significant bit is at index Self::BITS - 1

Examples

use bnum::types::I256;

let n = I256::from(0b001010100101010101u32);
assert!(n.bit(0));
assert!(!n.bit(1));
assert!(!n.bit(I256::BITS - 1));

pub const fn is_zero(&self) -> bool

Returns whether self is zero.

Examples

use bnum::types::I256;

assert!(I256::ZERO.is_zero());
assert!(!I256::ONE.is_zero());

pub const fn is_one(&self) -> bool

Returns whether self is one.

Examples

use bnum::types::I256;

assert!(I256::ONE.is_one());
assert!(!I256::MAX.is_one());

pub const fn from_bits(bits: BUintD32<N>) -> Self

Creates a signed integer with bits as its underlying representation in two’s complement.

This method is faster for casting from a BUintD32 to a BIntD32 of the same size than using the As trait.

pub const fn to_bits(self) -> BUintD32<N>

This simply returns the underlying representation of the integer in two’s complement, as an unsigned integer.

This method is faster for casting from a BIntD32 to a BUintD32 of the same size than using the As trait.